requirements.txt
google-cloud-bigquery==2.6.0
google-cloud-bigquery-storage==2.1.0
PyJWT==1.7.1
pandas
pyarrow